fix: moved PackSequencesTestActor to a to separate module to avoid pytest import in mcore worker venv
